+++ examples/es5/machine-learning/neuralNetwork.js
@@ -7,55 +7,28 @@ var mnist = require("./mnist");
 var ANN = require("./ann");
 var now = require("performance-now");
 
-var accuracy = async(regeneratorRuntime.mark(function _callee(af, predicted, target) {
-    var pMax, tMax;
+var accuracy = function accuracy(af, predicted, target) {
+    var pMax = af.findMaxAt(predicted, 1);
+    var tMax = af.findMaxAt(target, 1);
+    return 100 * af.count(pMax.index.eq(tMax.index)) / tMax.index.elements();
+};
+
+var annDemo = async(regeneratorRuntime.mark(function _callee(af, deviceInfo) {
+    var data, featureSize, trainFeats, testFeats, trainTarget, testTarget, network, start, end, trainOutput, testOutput;
     return regeneratorRuntime.wrap(function _callee$(_context) {
         while (1) {
             switch (_context.prev = _context.next) {
-                case 0:
-                    _context.next = 2;
-                    return af.findMaxAtAsync(predicted, 1);
-
-                case 2:
-                    pMax = _context.sent;
-                    _context.next = 5;
-                    return af.findMaxAtAsync(target, 1);
-
-                case 5:
-                    tMax = _context.sent;
-                    _context.next = 8;
-                    return af.countAsync(pMax.index.eq(tMax.index));
-
-                case 8:
-                    _context.t0 = _context.sent;
-                    _context.t1 = 100 * _context.t0;
-                    _context.t2 = tMax.index.elements();
-                    return _context.abrupt("return", _context.t1 / _context.t2);
-
-                case 12:
-                case "end":
-                    return _context.stop();
-            }
-        }
-    }, _callee, this);
-}));
-
-var annDemo = async(regeneratorRuntime.mark(function _callee2(af, deviceInfo) {
-    var data, featureSize, trainFeats, testFeats, trainTarget, testTarget, network, start, end, trainOutput, testOutput;
-    return regeneratorRuntime.wrap(function _callee2$(_context2) {
-        while (1) {
-            switch (_context2.prev = _context2.next) {
                 case 0:
                     console.log("Running ANN Demo on device:\n");
                     common.printDeviceInfo(deviceInfo);
                     console.log("");
 
                     console.log("Setting up training data.");
-                    _context2.next = 6;
+                    _context.next = 6;
                     return mnist.setup(af, true, 0.6);
 
                 case 6:
-                    data = _context2.sent;
+                    data = _context.sent;
                     featureSize = data.trainImages.elements() / data.numTrain;
 
                     // Reshape images into feature vectors
@@ -69,19 +42,17 @@ var annDemo = async(regeneratorRuntime.mark(function _callee2(af, deviceInfo) {
                     // Train network
 
                     start = now();
-                    _context2.next = 16;
-                    return network.train(trainFeats, trainTarget, {
+
+                    network.train(trainFeats, trainTarget, {
                         alpha: 1.0,
                         maxEpochs: 300,
                         batchSize: 100,
                         maxError: 0.0001
                     });
-
-                case 16:
-                    _context2.next = 18;
+                    _context.next = 17;
                     return af.waitAsync();
 
-                case 18:
+                case 17:
                     end = now();
 
                     // Run the trained network and test accuracy.
@@ -90,35 +61,19 @@ var annDemo = async(regeneratorRuntime.mark(function _callee2(af, deviceInfo) {
                     testOutput = network.predict(testFeats);
 
                     console.log("Training set:");
-                    _context2.t0 = console;
-                    _context2.next = 25;
-                    return accuracy(af, trainOutput, trainTarget);
-
-                case 25:
-                    _context2.t1 = _context2.sent.toFixed(2);
-                    _context2.t2 = "Accuracy on training data: " + _context2.t1;
-
-                    _context2.t0.log.call(_context2.t0, _context2.t2);
+                    console.log("Accuracy on training data: " + accuracy(af, trainOutput, trainTarget).toFixed(2));
 
                     console.log("Test set:");
-                    _context2.t3 = console;
-                    _context2.next = 32;
-                    return accuracy(af, testOutput, testTarget);
-
-                case 32:
-                    _context2.t4 = _context2.sent.toFixed(2);
-                    _context2.t5 = "Accuracy on testing  data: " + _context2.t4;
-
-                    _context2.t3.log.call(_context2.t3, _context2.t5);
+                    console.log("Accuracy on testing  data: " + accuracy(af, testOutput, testTarget).toFixed(2));
 
                     console.log("Training time: " + ((end - start) / 1000).toFixed(10) + " seconds\n");
 
-                case 36:
+                case 25:
                 case "end":
-                    return _context2.stop();
+                    return _context.stop();
             }
         }
-    }, _callee2, this);
+    }, _callee, this);
 }));
 
 common.runOnBestDevice(annDemo, "ANN Demo");