+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.21.0 (2015-12-30)
+
+- All calculation methods synchronus
+- T/H method added to AFArray
+
 # 0.20.0 (2015-12-28)
 
 - FIX: gfor actually works